pyproject.toml
[project]
name = "document-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "faiss-cpu>=1.12.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "openai>=1.106.1",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "wikipedia>=1.4.0",
]
